fix: rename lib to src to be compatible with ava
