+++ src/bot.js
