+++ src/bot.test.js
