+++ src/stringLength.js
